reorganized code with regions
